+++ js/validate.js
@@ -1,0 +1,3 @@
+const checkStringLength = (string, maxLength) => string.length < maxLength;
+
+export {checkStringLength};
